refactor: 구분자 Service 추상 클래스로 변경
- 구분자 Service 는 여러 가지 타입의 계산기에 쓰일 수 있음.
- 유연성을 높이기 위함.
